Purge now-unneeded boilerplate
TODO: make the definition of a Bales::Application result in the automatic creation of both a root Bales::Command and a subclass of Bales::Command::Help as a child of that root.

TODO extra credit: give Bales::Application the ability to set the params of the root command directly (particularly the `action` and `option`s, but also the `banner`, `summary`, and `description`) when being defined, thus removing any/all need to define the root command namespace entirely.